+++ tests/e2e/cases/dynamic_require_static_object_prototype_to_string/tag_Undefined.ts
@@ -1,0 +1,1 @@
+module.exports.label = "tag_Undefined";